Move Cypress utils out of shell
